+++ HexRowFaster/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.1.0
+
+- Now you can adjust rowing/backward speed on the fly via the BepInEx configuration manager.
+- Improved multiplayer compatibility by limiting changes to only the local player controlling the ship.
+
 ## v1.0.0
 
 - Initial release
+- Modify ship rowing/backward paddle force on all ships.